Use SocketCAN as data link type of CAN frame
In commit 93ca5ff, all CAN captures are changed to use Linux cooked
mode, which will have an additional 16-byte SLL header. This will
increase the size of captured frames saved in file, which is critical
for embedded environment. This commit restores to use SocketCAN as
the data link type of CAN frame.

Besides, the byte orders of CAN ID in SLL and SocketCAN are different.
This commit deals with the byte order swapping as well.